# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TacoCloud Public API
  version: 1.0.0
  description: |
    Contrato HTTP público v1. Las rutas equivalentes bajo `/api` son aliases
    temporales y responden con `Deprecation` y `Link` hacia su sucesor v1.
    La versión HTTP es independiente de la versión del contrato de eventos.
servers:
  - url: /
...
security:
  - basicAuth: []
paths:
  /api/v1/ingredients:
    get:
      tags: [Ingredients]
      summary: List public ingredients
      security: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Public ingredient catalog
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IngredientResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Ingredients]
      summary: Create an ingredient
      description: Requires ADMIN.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientRequest'
      responses:
        '201':
          description: Ingredient created
          headers:
            Location:
              description: URI of the created ingredient
              schema: {type: string, format: uri}
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientResponse'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/ingredients/{id}:
    parameters:
      - $ref: '#/components/parameters/IngredientId'
      - $ref: '#/components/parameters/CorrelationId'
    get:
      tags: [Ingredients]
      summary: Get an ingredient
      security: []
      responses:
        '200':
          description: Ingredient found
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/IngredientResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Ingredients]
      summary: Replace public ingredient attributes
      description: Requires ADMIN; body id must equal path id.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/IngredientRequest'}
      responses:
        '200':
          description: Ingredient updated
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/IngredientResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Ingredients]
      summary: Delete an ingredient
      description: Requires ADMIN.
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/admin/ingredients/{id}/catalog:
    parameters:
      - $ref: '#/components/parameters/IngredientId'
      - $ref: '#/components/parameters/CorrelationId'
    patch:
      tags: [Administration]
      summary: Update operational catalog fields
      description: Requires ADMIN and optimistic version.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/IngredientCatalogUpdateRequest'}
      responses:
        '200':
          description: Catalog updated
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/IngredientAdminResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /api/v1/admin/ingredients/{id}/stock-adjustments:
    parameters:
      - $ref: '#/components/parameters/IngredientId'
      - $ref: '#/components/parameters/CorrelationId'
    post:
      tags: [Administration]
      summary: Adjust ingredient stock
      description: Requires ADMIN and optimistic version.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/StockAdjustmentRequest'}
      responses:
        '200':
          description: Stock adjusted
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/IngredientAdminResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
...
  /api/v1/users/me/favorites:
    get:
      tags: [Favorites]
      summary: List the authenticated user's favorites
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Favorite taco page
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FavoritePageResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /api/v1/users/me/favorites/{tacoId}:
    parameters:
      - name: tacoId
        in: path
        required: true
        schema: {type: string}
      - $ref: '#/components/parameters/CorrelationId'
    put:
      tags: [Favorites]
      summary: Mark a taco as favorite
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Favorites]
      summary: Remove a taco from favorites
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /api/v1/admin/announcements:
    get:
      tags: [Administration]
      summary: List active operational announcements
      description: Requires ADMIN; expired announcements are excluded.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Active announcements without author data
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/AnnouncementResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Administration]
      summary: Create an operational announcement
      description: Requires ADMIN; author and creation time are server-owned.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AnnouncementRequest'}
      responses:
        '201':
          description: Announcement created
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AnnouncementResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/admin/announcements/{announcementId}:
    parameters:
      - name: announcementId
        in: path
        required: true
        schema: {type: string}
      - $ref: '#/components/parameters/CorrelationId'
    delete:
      tags: [Administration]
      summary: Delete an announcement by stable id
      description: Requires ADMIN.
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic credentials resolved by Spring Security. Browser form login may establish the same authenticated session.
  headers:
    CorrelationId:
      description: Validated request correlation or server-generated UUID.
      schema:
        type: string
        maxLength: 64
        pattern: '^[A-Za-z0-9._-]+$'
      example: client-operation-123
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Preserved when valid; otherwise the server generates a UUID and returns it in the response.
      schema:
        type: string
        maxLength: 64
        pattern: '^[A-Za-z0-9._-]+$'
      example: client-operation-123
...
    IngredientId:
      name: id
      in: path
      required: true
      schema: {type: string, minLength: 2, maxLength: 20}
...
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Size:
      name: size
      in: query
      schema: {type: integer, minimum: 1, maximum: 50, default: 20}
  responses:
    NoContent:
      description: Operation completed without a response body
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
    BadRequest:
      description: Invalid request or malformed JSON
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
          example:
            type: urn:tacocloud:problem:validation_error
            title: Invalid request
            status: 400
            detail: The request contains invalid fields.
            instance: /api/v1/orders
            code: VALIDATION_ERROR
            violations:
              - field: deliveryName
                reason: deliveryName is required
    Unauthorized:
      description: Authentication is required
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
    Forbidden:
      description: Authenticated principal lacks the required role or ownership
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
    NotFound:
      description: Resource not found or hidden by ownership rules
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
    Conflict:
      description: State, version, workflow, or idempotency conflict
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
    Unprocessable:
      description: Request is syntactically valid but violates a business rule
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
  schemas:
    IngredientType:
      type: string
      enum: [WRAP, PROTEIN, VEGGIES, CHEESE, SAUCE]
    DietaryTag:
      type: string
      enum: [VEGAN, VEGETARIAN, GLUTEN_FREE]
    Allergen:
      type: string
      enum: [GLUTEN, DAIRY, EGG, PEANUT, TREE_NUT, SOY]
    SpiceLevel:
      type: string
      enum: [NONE, MILD, MEDIUM, HOT, EXTREME]
...
    IngredientRequest:
      type: object
      additionalProperties: false
      required: [id, name, type, dietaryTags, allergens, spiceLevel]
      properties:
        id: {type: string, pattern: '^[A-Z0-9_-]{2,20}$', example: FLTO}
        name: {type: string, minLength: 1, maxLength: 50, example: Flour Tortilla}
        type: {$ref: '#/components/schemas/IngredientType'}
        dietaryTags:
          type: array
          uniqueItems: true
          items: {$ref: '#/components/schemas/DietaryTag'}
        allergens:
          type: array
          uniqueItems: true
          items: {$ref: '#/components/schemas/Allergen'}
        spiceLevel: {$ref: '#/components/schemas/SpiceLevel'}
    IngredientResponse:
      type: object
      additionalProperties: false
      required: [id, name, type, unitPrice, available, dietaryTags, allergens, spiceLevel]
      properties:
        id: {type: string}
        name: {type: string}
        type: {$ref: '#/components/schemas/IngredientType'}
        unitPrice: {type: number, format: decimal, minimum: 0, multipleOf: 0.01}
        available: {type: boolean}
        dietaryTags:
          type: array
          items: {$ref: '#/components/schemas/DietaryTag'}
        allergens:
          type: array
          items: {$ref: '#/components/schemas/Allergen'}
        spiceLevel: {$ref: '#/components/schemas/SpiceLevel'}
    IngredientCatalogUpdateRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        unitPrice: {type: number, format: decimal, minimum: 0, multipleOf: 0.01}
        available: {type: boolean}
        reorderLevel: {type: integer, minimum: 0}
        expectedVersion: {type: integer, format: int64, minimum: 0}
    StockAdjustmentRequest:
      type: object
      additionalProperties: false
      required: [quantity, expectedVersion]
      properties:
        quantity: {type: integer, format: int32}
        expectedVersion: {type: integer, format: int64, minimum: 0}
    IngredientAdminResponse:
      allOf:
        - $ref: '#/components/schemas/IngredientResponse'
        - type: object
          required: [stockOnHand, reorderLevel, version]
          properties:
            stockOnHand: {type: integer, minimum: 0}
            reorderLevel: {type: integer, minimum: 0}
            version: {type: integer, format: int64, minimum: 0}
...
    TacoClassificationResponse:
      type: object
      additionalProperties: false
      required: [tacoId, tacoName, dietaryTags, allergens, spiceLevel, disclaimer]
      properties:
        tacoId: {type: string, nullable: true}
        tacoName: {type: string}
        dietaryTags:
          type: array
          items: {$ref: '#/components/schemas/DietaryTag'}
        allergens:
          type: array
          items: {$ref: '#/components/schemas/Allergen'}
        spiceLevel: {$ref: '#/components/schemas/SpiceLevel'}
        disclaimer: {type: string}
    TacoCatalogResponse:
      type: object
      additionalProperties: false
      required: [id, name, createdAt, ingredients, classification]
      properties:
        id: {type: string, nullable: true}
        name: {type: string}
        createdAt: {type: string, format: date-time, nullable: true}
        ingredients:
          type: array
          items: {$ref: '#/components/schemas/IngredientResponse'}
        classification: {$ref: '#/components/schemas/TacoClassificationResponse'}
...
    FavoritePageResponse:
      type: object
      additionalProperties: false
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/TacoCatalogResponse'}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1}
        totalElements: {type: integer, format: int64, minimum: 0}
        totalPages: {type: integer, minimum: 0}
...
    AnnouncementSeverity:
      type: string
      enum: [INFO, WARNING, CRITICAL]
    AnnouncementRequest:
      type: object
      additionalProperties: false
      required: [text, severity, expiresAt]
      properties:
        text: {type: string, minLength: 1, maxLength: 500}
        severity: {$ref: '#/components/schemas/AnnouncementSeverity'}
        expiresAt: {type: string, format: date-time}
    AnnouncementResponse:
      type: object
      additionalProperties: false
      required: [id, text, severity, createdAt, expiresAt, active]
      properties:
        id: {type: string}
        text: {type: string}
        severity: {$ref: '#/components/schemas/AnnouncementSeverity'}
        createdAt: {type: string, format: date-time}
        expiresAt: {type: string, format: date-time}
        active: {type: boolean}
    ApiProblem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, violations]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        instance: {type: string}
        code: {type: string}
        violations:
          type: array
          items: {$ref: '#/components/schemas/Violation'}
    Violation:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field: {type: string}
        reason: {type: string}